refactor: split Remote and Local into their own files
